refactor: extract index.ts's tool_call handler into clients/runtime-tool-call.ts (closes #678)
Pure refactor, zero behavior change. tool_call was the one lifecycle hook
whose logic still lived inline in the entry point while its three siblings
(session_start, tool_result, turn_end/agent_end) had already been extracted;
it was also the largest (~800 lines) and most safety-critical, since it
mutates the agent's own edits (read-guard relocation/autopatch) before they
land.

New handleToolCall(deps) in clients/runtime-tool-call.ts follows the exact
same DI shape as its siblings; index.ts's pi.on("tool_call", ...) is now a
thin call site. Verified against the original line-by-line (diff of the
extracted body vs. pre-extraction source shows only necessary renames/import
fixes) plus the full existing test suite passing unmodified.

Two of tests/index-integration.test.ts's tool_call tests needed their
../clients/lsp/index.js mock completed with a resetLSPService stub: the new
call site always forwards resetLSPService into the DI object (matching how
tool_result's own call site already does this), so it's now read on every
tool_call instead of only deep inside the rarely-hit partial-apply branch.
No functional change — resetLSPService is a real module export in
production, only these tests' partial mock was incomplete against it.

New tests/clients/runtime-tool-call.test.ts unit-tests handleToolCall
directly against mocked deps.
